+++ src/extra/components/__init__.py
@@ -1,3 +1,4 @@
 
 from .scantool import Scantool  # noqa
 from .pulses import XrayPulses, OpticalLaserPulses, DldPulses  # noqa
+from .scan import Scan